requirements.txt
pyindi-client
astropy
#astropy[all]
opencv-python
ephem
parimiko
